refactor(importer): extract import services to internal/importer/ package (PKG-4b)
Move import_service.go and import_path_service.go from internal/server/ to
internal/importer/. Replace *Server receiver/field accesses with explicit
constructor parameters. Update all call sites in internal/server/.

Files moved:
- internal/server/import_service.go -> internal/importer/service.go
- internal/server/import_path_service.go -> internal/importer/path_service.go
- internal/server/import_service_test.go -> internal/importer/service_test.go
- internal/server/import_path_service_test.go -> internal/importer/path_service_test.go

Files updated:
- internal/server/server.go: field types + constructors use importer.*
- internal/server/deluge_discovery.go: ImportFileRequest -> importer.ImportFileRequest
- internal/server/filesystem_handlers.go: ImportFileRequest -> importer.ImportFileRequest
- internal/server/service_layer_test.go: NewImportPathService/ErrImportPathEmpty -> importer.*

*Server dependencies resolved:
- No *Server receiver ever existed on these services; they were already
  standalone struct types. The only coupling was package-level (same package
  as server). Resolved by moving to a new package with explicit constructor params.
- CreateIngestVersion/IngestVersionParams: were thin wrappers in server/
  over internal/versions; replaced with direct versions.CreateIngestVersion
  and versions.IngestVersionParams imports.
- stringPtr helper: was in server_helpers.go; added a local copy in service.go.

Refs: PKG-4b
